masp_agent/requirements.txt
google-generativeai
python-socketio
requests
python-dotenv